backend/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
joblib==1.3.2
python-dotenv==1.0.1
pytest==7.4.4
httpx==0.27.2
matplotlib==3.8.2
torch==2.2.1
transformers==4.38.2
xgboost==2.0.3
Pillow==10.2.0
beautifulsoup4==4.12.3
timm==0.9.16
SQLAlchemy==2.0.29
alembic==1.13.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
redis==5.0.3
